fix: cleanup code (#1166)
- move revision tests to own file
- separate generate_data_mixin
- fix test cleanup for test_publish
- remove unused code from Dataset
- Use tempfile to cleanup test data.
- add s3_bucket to store bucket resources.
