+++ tests/unit/backend/corpora/common/entities/test_dataset.py
@@ -1,3 +1,5 @@
+import tempfile
+
 import filecmp
 import os
 import typing
@@ -334,19 +336,18 @@ class TestDataset(CorporaTestCaseUsingMockAWS):
             description="describe another geneset",
             genes=genes1,
         )
-
-        csv_file = dataset.generate_tidy_csv_for_all_linked_genesets()
-        expected_csv = os.path.abspath(
-            os.path.join(os.path.dirname(__file__), "../../../fixtures/sample_geneset_csv.csv")
-        )  # noqa
-        self.assertTrue(filecmp.cmp(csv_file, expected_csv, shallow=False))
+        with tempfile.TemporaryDirectory() as temp_dir_name:
+            csv_file = dataset.generate_tidy_csv_for_all_linked_genesets(temp_dir_name)
+            expected_csv = os.path.abspath(
+                os.path.join(os.path.dirname(__file__), "../../../fixtures/sample_geneset_csv.csv")
+            )  # noqa
+            self.assertTrue(filecmp.cmp(csv_file, expected_csv, shallow=False))
 
     def test__generate_tidy_csv_for_all_linked_genesets__stores_file_in_correct_location(self):
         collection = self.generate_collection(self.session)
-        deployment_directory_params = dict(url="test_url")
-        dataset = self.generate_dataset(
-            self.session, collection=collection, deployment_directories=[deployment_directory_params]
-        )
+        dataset = self.generate_dataset(self.session, collection=collection)
+        deployment_directory_params = dict(url=f"http://test_domain.example/e/{dataset.id}.cxg/")
+        dataset.update(deployment_directories=[deployment_directory_params])
         genes0 = [
             {"gene_symbol": "HBB", "gene_description": "gene 1", "additional_params": {}},
             {"gene_symbol": "HBA1", "gene_description": "gene 2", "additional_params": {}},
@@ -365,10 +366,13 @@ class TestDataset(CorporaTestCaseUsingMockAWS):
             description="describe the geneset",
             genes=genes0,
         )
-        csv_file = dataset.generate_tidy_csv_for_all_linked_genesets()
-        dataset.copy_csv_to_s3(csv_file)
+        with tempfile.TemporaryDirectory() as temp_dir_name:
+            csv_file = dataset.generate_tidy_csv_for_all_linked_genesets(temp_dir_name)
+            s3_file = dataset.copy_csv_to_s3(csv_file)
+        expected_suffix = f"{dataset.id}-genesets.csv"
+        self.assertTrue(s3_file.endswith(expected_suffix), msg=f"{s3_file} does not end with {expected_suffix}")
         stored_files = [x.key for x in self.cellxgene_bucket.objects.all()]
-        self.assertIn(csv_file, stored_files)
+        self.assertIn(s3_file, stored_files)
 
     def assertRowsDeleted(self, tests: typing.List[typing.Tuple[str, Base]]):
         """
